+++ .github/dependabot.yml
@@ -26,3 +26,14 @@ updates:
       - dependency-type: "production"
     commit-message:
       prefix: "gradle"
+
+  - package-ecosystem: "gradle"
+    directory: "/ihub-spring/"
+    schedule:
+      interval: "daily"
+    assignees:
+      - "henry-hub"
+    allow:
+      - dependency-type: "production"
+    commit-message:
+      prefix: "gradle"
