+++ ihub-plugins/src/test/groovy/pub/ihub/plugin/IHubBasicPluginTest.groovy
@@ -18,7 +18,7 @@ package pub.ihub.plugin
 import groovy.util.logging.Slf4j
 import org.gradle.api.Project
 import org.gradle.testfixtures.ProjectBuilder
-import spock.lang.Ignore
+import pub.ihub.plugin.test.IHubSpecification
 import spock.lang.Title
 
 import static org.gradle.api.initialization.Settings.DEFAULT_SETTINGS_FILE
@@ -269,10 +269,6 @@ iHub.repoIncludeGroupRegex=pub\\.ihub\\..*
         result.output.contains 'BUILD SUCCESSFUL'
     }
 
-    /**
-     * TODO 独立SpringBoot插件
-     */
-    @Ignore
     def 'Java插件配置测试'() {
         setup: '初始化项目（此处借用子项目测试，主项目不触发beforeEvaluate）'
         copyProject 'basic.gradle'
@@ -424,38 +420,6 @@ iHubPublish.publishDocs=true
 
         when: '构建项目'
         def result = gradleBuilder.build()
-
-        then: '检查结果'
-        result.output.contains 'BUILD SUCCESSFUL'
-    }
-
-    /**
-     * TODO 更换Native插件
-     */
-    @Ignore
-    def 'Native插件配置测试'() {
-        setup: '初始化项目'
-        copyProject 'basic.gradle'
-        buildFile << '''
-            apply {
-                plugin 'pub.ihub.plugin.ihub-native'
-            }
-        '''
-
-        when: '构建项目'
-        def result = gradleBuilder.build()
-
-        then: '检查结果'
-        result.output.contains 'BUILD SUCCESSFUL'
-
-        when: '构建项目'
-        buildFile << '''
-            iHubNative {
-                bpJvmVersion = '11'
-            }
-        '''
-        testProjectDir.newFile('.java-local.properties') << 'spring.profiles.active=dev'
-        result = gradleBuilder.build()
 
         then: '检查结果'
         result.output.contains 'BUILD SUCCESSFUL'
