+++ ihub-plugins/src/test/groovy/pub/ihub/plugin/IHubBomPluginTest.groovy
@@ -16,6 +16,7 @@
 package pub.ihub.plugin
 
 import groovy.util.logging.Slf4j
+import pub.ihub.plugin.test.IHubSpecification
 import spock.lang.Title
 
 import static org.gradle.api.Project.DEFAULT_BUILD_FILE
