+++ ihub-spring/build.gradle
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2021 Henry 李恒 (henry.box@outlook.com).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+description = 'IHub Spring Gradle Plugins'
+
+repositories {
+    gradlePluginPortal()
+}
+
+dependencies {
+    implementation project(':ihub-aware'), project(':ihub-plugins')
+    implementation 'org.springframework.boot:spring-boot-gradle-plugin:3.0.0'
+    compileOnly    'org.springframework.boot:spring-boot-buildpack-platform'
+    implementation 'org.graalvm.buildtools:native-gradle-plugin:0.9.18'
+
+    testImplementation project(':ihub-test')
+}
+
+gradlePlugin {
+    plugins {
+        iHubBoot {
+            id                  = 'pub.ihub.plugin.ihub-boot'
+            displayName         = 'IHub Boot Gradle Plugin'
+            description         = 'IHub Boot Gradle Plugin'
+            implementationClass = 'pub.ihub.plugin.spring.IHubBootPlugin'
+        }
+        iHubNative {
+            id                  = 'pub.ihub.plugin.ihub-native'
+            displayName         = 'IHub Native Gradle Plugin'
+            description         = 'IHub Native Gradle Plugin'
+            implementationClass = 'pub.ihub.plugin.spring.IHubNativePlugin'
+        }
+    }
+}
