+++ ihub-spring/src/main/groovy/pub/ihub/plugin/spring/IHubNativePlugin.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Henry 李恒 (henry.box@outlook.com).
+ * Copyright (c) 2022 Henry 李恒 (henry.box@outlook.com).
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,8 +15,7 @@
  */
 package pub.ihub.plugin.spring
 
-import org.springframework.aot.gradle.SpringAotGradlePlugin
-import org.springframework.aot.gradle.dsl.SpringAotExtension
+import org.graalvm.buildtools.gradle.NativeImagePlugin
 import org.springframework.boot.gradle.tasks.bundling.BootBuildImage
 import pub.ihub.plugin.IHubPlugin
 import pub.ihub.plugin.IHubProjectPluginAware
@@ -30,7 +29,7 @@ import static pub.ihub.plugin.IHubProjectPluginAware.EvaluateStage.AFTER
  * 参考官方入门文档：https://docs.spring.io/spring-native/docs/current/reference/htmlsingle/
  * @author henry
  */
-@IHubPlugin(value = IHubNativeExtension, beforeApplyPlugins = [IHubBootPlugin, SpringAotGradlePlugin])
+@IHubPlugin(value = IHubNativeExtension, beforeApplyPlugins = [IHubBootPlugin, NativeImagePlugin])
 class IHubNativePlugin extends IHubProjectPluginAware<IHubNativeExtension> {
 
     @Override
@@ -38,16 +37,7 @@ class IHubNativePlugin extends IHubProjectPluginAware<IHubNativeExtension> {
         withExtension(AFTER) { ext ->
             withTask(BootBuildImage) {
                 it.builder = 'paketobuildpacks/builder:tiny'
-                it.environment = ext.environment
-            }
-            withExtension(SpringAotExtension) {
-                it.mode.set ext.aotMode
-                it.debugVerify.set ext.aotDebugVerify
-                it.removeXmlSupport.set ext.aotRemoveXmlSupport
-                it.removeSpelSupport.set ext.aotRemoveSpelSupport
-                it.removeYamlSupport.set ext.aotRemoveYamlSupport
-                it.removeJmxSupport.set ext.aotRemoveJmxSupport
-                it.verify.set ext.aotVerify
+                it.environment = withExtension(IHubBootExtension).environment + ext.environment
             }
         }
     }
