+++ ihub-test/build.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Henry 李恒 (henry.box@outlook.com).
+ * Copyright (c) 2022 Henry 李恒 (henry.box@outlook.com).
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-plugins {
-    id 'pub.ihub.plugin.ihub-settings' version '1.2.11'
-}
+description = 'IHub Gradle Plugins Test'
 
-iHubSettings {
-    includeProjects 'ihub-aware', 'ihub-plugins', 'ihub-spring', 'ihub-settings', 'ihub-test' noPrefix
+dependencies {
+    implementation gradleTestKit()
+    implementation 'org.spockframework:spock-core'
 }
